remote, librería del ciclo de vida de liveData, dataRepository y viewModel

+++ app/src/main/java/com/example/tapas_diego_unai/domain/Appetizer.kt
@@ -1,3 +1,3 @@
 package com.example.tapas_diego_unai.domain
 
-data class Appetizer(val ranking: String,val name: String, val location: String, val totalPoints: Int, val averagePoints: Int)
+data class Appetizer(val ranking: String,val name: String, val location: String, val totalPoints: Int, val averagePoints: Double)
